study_dir/Hojun/image/requirements.txt
alembic==1.13.1
banal==1.0.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
colorama==0.4.6
dataset==1.6.2
greenlet==3.0.3
idna==3.7
Mako==1.3.3
MarkupSafe==2.1.5
pillow==10.3.0
pluggy==1.5.0
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.52
sqlite-fts4==1.0.3
sqlite-utils==3.36
tabulate==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
watchdog==4.0.0
